Introduced notion of project "family".
In a multi-module project, all (transitive) children of the top-most parent are
a family. When searching for a symbol, we search a project's family as well as
its dependencies (but as with depends, we don't search the depends of our
family members).

This allows one to easily jump around within a family of modules, which one is
often wont to do.