Put all codex metadata in .codex instead of scattering it about.
Though it is nice to think that you can delete a directory and its codex
metadata goes away as well, but it's both a) annoying to scatter crap all over
the file system, and b) some projects have the same "root" (DLL projects), and
c) some project roots are on non-writable file systems (like when we're
grinding MonoTouch standard libraries or the Java source from java.home).

I'll eventually add some sort of project garbage collector that blows away
projects and metadata for projects whose root disappers and we'll call it good.